GeniusYield.CardanoApi.EraHistory

Description

Documentation

extractEraSummaries ∷ EraHistory CardanoMode → Summary (CardanoEras StandardCrypto)

Extract the Summary from Cardano EraHistory.

This is safe as long as Ouroboros.Interpeter is defined as a newtype to Ouroboros.Summary.

TODO: Is there a way to ensure, at runtime, that Interpreter is still a newtype to Summary?

showEraSummaries ∷ EraHistory CardanoMode → String

Extract and show the era summaries with all its details, useful for manually constructing EraHistory.

See: "GeniusYield.Providers.Common.preprodEraHist"

getEraEndSlot ∷ EraHistory CardanoMode → Maybe SlotNo

Get the slot after which the current era ends.