GeniusYield.HTTP.Errors

Description

Documentation

class IsGYApiError e where

Class of types that can be converted into an HTTP API error.

Minimal complete definition

Nothing

Methods

toApiError ∷ e → GYApiError

default toApiError ∷ Exception e ⇒ e → GYApiError

data GYApiError

An example error code can be: INSUFFICIENT_BALANCE (i.e. it is not the HTTP status error message)

The message can be any textual representation of the error with more information.

The status code should be the HTTP status code.

Constructors

GYApiError

Fields

gaeErrorCode ∷ Text

gaeHttpStatus ∷ Status

gaeMsg ∷ Text

someBackendError ∷ Text → GYApiError

Create a typical BACKEND_ERROR internal serval error with given message.