GeniusYield.Providers.Common

Description

Documentation

data SomeDeserializeError

Constructors

DeserializeErrorBech32 Bech32DecodeError

DeserializeErrorAeson Text

DeserializeErrorAssetClass Text

DeserializeErrorScriptDataJson ScriptDataJsonError

DeserializeErrorHex Text

DeserializeErrorAddress

newServantClientEnv ∷ String → IO ClientEnv

Creates a new Servant ClientEnv from a base url.

parseEraHist ∷ (t → EraSummary) → [t] → Maybe (EraHistory CardanoMode)

Convert a regular list of era summaries (a la Ogmios) into a typed EraHistory (a la Ouroboros).

NOTE ==

This must be updated with each hardfork.

See NonEmptyConsSummaryCardanoEras types to understand why one cannot trivially automate this.

Well, unless one uses vectors, from dependent type land.

babbageProtocolVersion ∷ Natural

Hardcoded babbage protocol Version

preprodEraHist ∷ Interpreter (CardanoEras StandardCrypto)

Hardcoded era history for preprod. FIXME: Remove this hack; https://github.com/geniusyield/Core/issues/433

See: "GeniusYield.CardanoApi.EraHistory.showEraHistory"

previewEraHist ∷ Interpreter (CardanoEras StandardCrypto)

mainnetEraHist ∷ Interpreter (CardanoEras StandardCrypto)

silenceHeadersClientError ∷ ClientError → ClientError

Remove request headers info from returned ClientError.

This is used as quick and simple way to hide confidential information such as API token.