GeniusYield.Providers.LiteChainIndex

Description

Documentation

data LCIClient

A very simple chain index client, which only maintains a datum hashes.

withLCIClient

Arguments

∷ LocalNodeConnectInfo CardanoMode

→ [ChainPoint]
resume points

→ (LCIClient → IO r)

→ IO r

newLCIClient

Arguments

∷ LocalNodeConnectInfo CardanoMode

→ [ChainPoint]
resume points

→ IO LCIClient

Create new LCIClient.

Use withLCIClient if possible.

closeLCIClient ∷ LCIClient → IO ()

Close (destroy) LCIClient.

lciWaitUntilSlot ∷ LCIClient → GYSlot → IO GYSlot

Wait until LCIClient has processed a given slot.

lciLookupDatum ∷ LCIClient → GYLookupDatum

lciGetCurrentSlot ∷ LCIClient → IO GYSlot

This is not good gyGetCurrentSlot provider as it might lag plenty behind the current slot of local node.

lciStats ∷ LCIClient → IO (GYSlot, Int)

Return statistics of LCIClient: currently processed slot and number of hashes known.