GeniusYield.Test.Privnet.Ctx

Description

Context

data Ctx

Constructors

Ctx

Fields

ctxEra ∷ !GYEra

ctxInfo ∷ !(LocalNodeConnectInfo CardanoMode)

ctxLCI ∷ !LCIClient

ctxDbSync ∷ !(Maybe CardanoDbSyncConn)

ctxUser1 ∷ !User

ctxUser2 ∷ !User

ctxUser3 ∷ !User

ctxGold ∷ !GYAssetClass

asset used in tests

ctxIron ∷ !GYAssetClass

asset used in tests

ctxLog ∷ !GYLog

ctxLookupDatum ∷ !GYLookupDatum

ctxQueryUtxos ∷ !GYQueryUTxO

ctxGetParams ∷ !GYGetParameters

User

data User

Constructors

User

Fields

userSKey ∷ !GYPaymentSigningKey

userAddr ∷ !GYAddress

userColl ∷ !GYTxOutRef

data UserIdx

Constructors

User1
user 1 is used as geniusyield user

User2

User3

userPkh ∷ User → GYPubKeyHash

userVKey ∷ User → GYPaymentVerificationKey

Operations

ctxRunI ∷ Ctx → UserIdx → GYTxMonadNode (GYTxSkeleton v) → IO GYTxBody

ctxRunC ∷ ∀ a. Ctx → UserIdx → GYTxMonadNode a → IO a

ctxRunF ∷ ∀ t v. Traversable t ⇒ Ctx → UserIdx → GYTxMonadNode (t (GYTxSkeleton v)) → IO (t GYTxBody)

__ctxRunF ∷ ∀ t v. Traversable t ⇒ Ctx → User → GYTxMonadNode (t (GYTxSkeleton v)) → IO (t GYTxBody)

ctxCurrentSlot ∷ Ctx → IO GYSlot

ctxWaitNextSlot ∷ Ctx → IO ()

ctxWaitUntilSlot ∷ Ctx → GYSlot → IO ()

ctxProviders ∷ Ctx → GYProviders

ctxSlotConfig ∷ Ctx → IO GYSlotConfig

submitTx ∷ Ctx → UserIdx → GYTxBody → IO GYTxId

__submitTx ∷ Ctx → User → GYTxBody → IO GYTxId

Helpers

ctxUser ∷ Ctx → UserIdx → User

ctxUserAddr ∷ Ctx → UserIdx → GYAddress

ctxUserPkh ∷ Ctx → UserIdx → GYPubKeyHash

newTempUserCtx

Arguments

∷ Ctx

→ UserIdx
User which will fund this new user.

→ GYValue
Describes balance of new user.

→ IO User

Creates a new user with the given balance. Note that we'll deduct 5 ada from the given fund as a collateral output.

ctxQueryBalance ∷ Ctx → UserIdx → IO GYValue

findOutput ∷ GYAddress → GYTxBody → IO GYTxOutRef

Function to find for the first locked output in the given GYTxBody at the given GYAddress.

addRefScriptCtx

Arguments

∷ Ctx
Given context.

→ UserIdx
User which will execute the transaction (if required).

→ GYScript 'PlutusV2
Given script.

→ IO GYTxOutRef
Returns the reference for the desired output.

Function to add for a reference script. It adds the script in so called "Always failing" validator so that it won't be later possible to spend this output. There is a slight optimisation here in that if the desired reference script already exists then we don't add another one and return the reference for the found one else, we create a new one.

addRefInputCtx

Arguments

∷ Ctx
Given context.

→ UserIdx
User which will execute this transaction.

→ Bool
Whether to inline the datum.

→ GYAddress
Address to put this output at.

→ GYDatum
The datum to put.

→ IO GYTxOutRef
Returns the reference for the required output.

Function to add for a reference input.