GeniusYield.Transaction.CoinSelection

Description

Documentation

data GYBalancedTx v

An *almost* finalized Tx.

This is fully balanced _except_ potentially missing an ada change output, and missing the exact fee. Both of these will be set by finalizeGYBalancedTx.

Constructors

GYBalancedTx

Fields

gybtxIns ∷ ![GYTxInDetailed v]

gybtxCollaterals ∷ !GYUTxOs

gybtxOuts ∷ ![GYTxOut v]

gybtxMint ∷ !(Maybe (GYValue, [(Some GYMintingPolicy, GYRedeemer)]))

gybtxInvalidBefore ∷ !(Maybe GYSlot)

gybtxInvalidAfter ∷ !(Maybe GYSlot)

gybtxSigners ∷ !(Set GYPubKeyHash)

gybtxRefIns ∷ !GYUTxOs

data GYTxInDetailed v

A further detailed version of GYTxIn, containing all information about a UTxO.

Constructors

GYTxInDetailed

Fields

gyTxInDet ∷ GYTxIn v

gyTxInDetAddress ∷ GYAddress

gyTxInDetValue ∷ GYValue

gyTxInDetScriptRef ∷ Maybe (Some GYScript)

gyTxInDetInlineDat ∷ Bool

Did GYUtXo had inline datum?

data GYCoinSelectionEnv v

Constructors

GYCoinSelectionEnv

Fields

existingInputs ∷ ![GYTxInDetailed v]

List of existing inputs that must be used.

requiredOutputs ∷ ![(GYAddress, GYValue)]

Outputs to pay for.

mintValue ∷ !GYValue

Value minted in the transaction.

changeAddr ∷ GYAddress

Address where change outputs will be sent to.

ownUtxos ∷ !GYUTxOs

Set of own utxos to select additional inputs from.

extraLovelace ∷ !Natural

Extra lovelace to look for on top of outputs (for fee).

minimumUTxOF ∷ GYTxOut v → Natural

maxValueSize ∷ Natural

data GYCoinSelectionStrategy

Constructors

GYLargestFirstMultiAsset

GYRandomImproveMultiAsset

GYLegacy

selectInputs ∷ ∀ m v. MonadRandom m ⇒ GYCoinSelectionEnv v → GYCoinSelectionStrategy → ExceptT BalancingError m ([GYTxInDetailed v], [GYTxOut v])

Select additional inputs from the set of own utxos given, such that when combined with given existing inputs, they cover for all the given outputs, as well as extraLovelace.

Return the list of additional inputs chosen and the change outputs created.

NOTE: This function doesn't necessarily take care of all ada change. Indeed, if any extra ada is still in the inputs and doesn't appear in the change outputs selected - it will simply be taken care of by makeTransactionBodyAutoBalance called higher in the call hierarchy.

Invariant ==

The ownUtxos and requiredOutputs arguments passed must contain non negative GYValues, with each one containing a positive amount of ada.