GeniusYield.TxBuilder

Description

Documentation

module GeniusYield.TxBuilder.Run

module GeniusYield.TxBuilder.Random

module GeniusYield.TxBuilder.NodeQuery

module GeniusYield.TxBuilder.Node

module GeniusYield.TxBuilder.Errors

module GeniusYield.TxBuilder.Common

module GeniusYield.TxBuilder.Class

queryBalance ∷ GYTxQueryMonad m ⇒ GYAddress → m GYValue

Query the balance at given address.

getAdaOnlyUTxO ∷ GYTxQueryMonad m ⇒ GYAddress → m [(GYTxOutRef, Natural)]

Query the txoutrefs at given address with ADA-only values.

Useful for finding a txoutref to be used as collateral.

adaOnlyUTxOPure ∷ GYUTxOs → [(GYTxOutRef, Natural)]

getCollateral'

Arguments

∷ GYTxQueryMonad m

⇒ GYAddress
The address where to look.

→ Natural
The minimal amount of lovelace required as collateral.

→ m (Maybe (GYTxOutRef, Natural))
Returns the smallest ada-only UTxO and the contained amount of lovelace at the specified address with the specified minimal value. If no such UTxO exists, Nothing is returned.

Get a UTxO suitable for use as collateral.

getCollateral

Arguments

∷ GYTxQueryMonad m

⇒ GYAddress
The address where to look.

→ Natural
The minimal amount of lovelace required as collateral.

→ m (GYTxOutRef, Natural)
Returns the smallest ada-only UTxO and the contained amount of lovelace at the specified address with the specified minimal value. If no such UTxO exists, an exception is thrown.

Get an UTxO suitable for use as collateral.

getTxBalance ∷ GYTxQueryMonad m ⇒ GYPubKeyHash → GYTx → m GYValue

Calculate how much balance is the given transaction is moving to given pubkeyhash address(es).