GeniusYield.TxBuilder.Class

Description

Documentation

class Monad m ⇒ MonadError e (m ∷ Type → Type) | m → e where

The strategy of combining computations that can throw exceptions by bypassing bound functions from the point an exception is thrown to the point that it is handled.

Is parameterized over the type of error information and the monad type constructor. It is common to use Either String as the monad type constructor for an error monad in which error descriptions take the form of strings. In that case and many other common cases the resulting monad is already defined as an instance of the MonadError class. You can also define your own error type and/or use a monad type constructor other than Either String or Either IOError. In these cases you will have to explicitly define instances of the MonadError class. (If you are using the deprecated Control.Monad.Error or Control.Monad.Trans.Error, you may also have to define an Error instance.)

Methods

throwError ∷ e → m a

Is used within a monadic computation to begin exception processing.

catchError ∷ m a → (e → m a) → m a

A handler function to handle previous errors and return to normal execution. A common idiom is:

do { action1; action2; action3 } `catchError` handler

where the action functions can call throwError. Note that handler and the do-block must have the same return type.

class Monad m ⇒ MonadRandom (m ∷ Type → Type) where

Methods

getRandomR ∷ Random a ⇒ (a, a) → m a

getRandom ∷ Random a ⇒ m a

getRandomRs ∷ Random a ⇒ (a, a) → m [a]

getRandoms ∷ Random a ⇒ m [a]

class GYTxQueryMonad m ⇒ GYTxMonad m where

Class of monads for querying monads as a user.

Methods

someUTxO ∷ m GYTxOutRef

Return some unspend transaction output

Note: may or may not return the same value

randSeed ∷ m Int

A seed to inject non-determinism.

class MonadError GYTxMonadException m ⇒ GYTxQueryMonad m where

Class of monads for querying chain data.

Minimal complete definition

networkId, lookupDatum, (utxoAtTxOutRef | utxosAtTxOutRefs), (utxosAtAddress | utxosAtAddresses), slotConfig, currentSlot, logMsg

Methods

networkId ∷ m GYNetworkId

Get the network id

lookupDatum ∷ GYDatumHash → m (Maybe GYDatum)

Lookup datum by its hash.

utxoAtTxOutRef ∷ GYTxOutRef → m (Maybe GYUTxO)

Lookup GYUTxO at GYTxOutRef.

utxosAtTxOutRefs ∷ [GYTxOutRef] → m GYUTxOs

Lookup GYUTxOs at multiple GYTxOutRefs at once

utxosAtAddress ∷ GYAddress → m GYUTxOs

Lookup GYUTxOs at GYAddress.

utxosAtAddresses ∷ [GYAddress] → m GYUTxOs

Lookup GYUTxOs at zero or more GYAddress.

utxoRefsAtAddress ∷ GYAddress → m [GYTxOutRef]

Lookup the `[GYTxOutRef]`s at a GYAddress

slotConfig ∷ m GYSlotConfig

Obtain the slot config for the network.

Implementations using era history to create slot config may raise GYEraSummariesToSlotConfigError.

currentSlot ∷ m GYSlot

Lookup the current GYSlot.

logMsg ∷ HasCallStack ⇒ GYLogNamespace → GYLogSeverity → String → m ()

Log a message with specified namespace and severity.

data GYTxSkeleton (v ∷ PlutusVersion)

Transaction skeleton

Note: let's add fields as we need them.

The parameter v indicates the minimum version of scripts allowed as inputs.

Constructors

GYTxSkeleton

Fields

gytxIns ∷ ![GYTxIn v]

gytxOuts ∷ ![GYTxOut v]

gytxRefIns ∷ !(GYTxSkeletonRefIns v)

gytxMint ∷ !(Map (Some GYMintingPolicy) (Map GYTokenName Integer, GYRedeemer))

gytxSigs ∷ !(Set GYPubKeyHash)

gytxInvalidBefore ∷ !(Maybe GYSlot)

gytxInvalidAfter ∷ !(Maybe GYSlot)

data GYTxSkeletonRefIns ∷ PlutusVersion → Type where

Constructors

GYTxSkeletonRefIns ∷ VersionIsGreaterOrEqual v PlutusV2 ⇒ !(Set GYTxOutRef) → GYTxSkeletonRefIns v

GYTxSkeletonNoRefIns ∷ GYTxSkeletonRefIns v

gyTxSkeletonRefInsToList ∷ GYTxSkeletonRefIns v → [GYTxOutRef]

data RandT g (m ∷ Type → Type) a

lookupDatum' ∷ GYTxQueryMonad m ⇒ GYDatumHash → m GYDatum

A version of lookupDatum that raises GYNoDatumForHash if the datum is not found.

utxoAtTxOutRef' ∷ GYTxQueryMonad m ⇒ GYTxOutRef → m GYUTxO

A version of utxoAtTxOutRef that raises GYNoUtxoAtRef if the utxo is not found.

slotToBeginTime ∷ GYTxQueryMonad f ⇒ GYSlot → f GYTime

Get the starting GYTime of a GYSlot in GYTxMonad.

slotToEndTime ∷ GYTxQueryMonad f ⇒ GYSlot → f GYTime

Get the ending GYTime of a GYSlot (inclusive) in GYTxMonad.

enclosingSlotFromTime ∷ GYTxQueryMonad f ⇒ GYTime → f (Maybe GYSlot)

Get the GYSlot of a GYTime in GYTxMonad.

Returns Nothing if given time is before known system start.

enclosingSlotFromTime' ∷ GYTxQueryMonad m ⇒ GYTime → m GYSlot

Partial version of enclosingSlotFromTime.

Raises GYTimeUnderflowException if given time is before known system start.

scriptAddress ∷ GYTxQueryMonad m ⇒ GYValidator v → m GYAddress

Calculate script's address.

scriptAddress' ∷ GYTxQueryMonad m ⇒ GYValidatorHash → m GYAddress

Calculate script's address.

addressFromText' ∷ MonadError GYTxMonadException m ⇒ Text → m GYAddress

Parse the bech32 representation of an address into GYAddress in GYTxMonad.

Throw GYConversionException if parsing fails.

addressFromPlutusM ∷ GYTxQueryMonad m ⇒ Address → m (Either PlutusToCardanoError GYAddress)

Convert a Address to GYAddress in GYTxMonad.

Explicitly returns an error rather than throwing it.

addressFromPlutusHushedM ∷ GYTxQueryMonad m ⇒ Address → m (Maybe GYAddress)

hushed version of addressFromPlutusM.

addressFromPlutus' ∷ GYTxQueryMonad m ⇒ Address → m GYAddress

Convert a Address to GYAddress in GYTxMonad.

Throw GYConversionException if conversion fails.

addressToPubKeyHash' ∷ MonadError GYTxMonadException m ⇒ GYAddress → m GYPubKeyHash

Convert GYAddress to GYPubKeyHash in GYTxMonad.

Throw GYConversionException if address is not key-hash one.

addressToPubKeyHashIO ∷ GYAddress → IO GYPubKeyHash

addressToValidatorHash' ∷ MonadError GYTxMonadException m ⇒ GYAddress → m GYValidatorHash

Convert GYAddress to GYValidatorHash in GYTxMonad.

Throw GYConversionException if address is not script-hash one.

addressToValidatorHashIO ∷ GYAddress → IO GYValidatorHash

valueFromPlutus' ∷ MonadError GYTxMonadException m ⇒ Value → m GYValue

Convert a Value to GYValue in GYTxMonad.

Throw GYConversionException if conversion fails.

valueFromPlutusIO ∷ Value → IO GYValue

Convert a Value to GYValue in IO.

Throw GYConversionException if conversion fails.

makeAssetClass' ∷ MonadError GYTxMonadException m ⇒ Text → Text → m GYAssetClass

Create a GYAssetClass from the textual representation of currency symbol and token name in GYTxMonad.

Throw GYConversionException if conversion fails.

makeAssetClassIO ∷ Text → Text → IO GYAssetClass

makeAssetClass' in the IO monad.

Throw GYConversionException if conversion fails.

assetClassFromPlutus' ∷ MonadError GYTxMonadException m ⇒ AssetClass → m GYAssetClass

Convert a AssetClass to GYAssetClass in GYTxMonad.

Throw GYConversionException if conversion fails.

tokenNameFromPlutus' ∷ MonadError GYTxMonadException m ⇒ TokenName → m GYTokenName

Convert a TokenName to GYTokenName in GYTxMonad.

Throw GYConversionException if conversion fails.

txOutRefFromPlutus' ∷ MonadError GYTxMonadException m ⇒ TxOutRef → m GYTxOutRef

Convert a TxOutRef to GYTxOutRef in GYTxMonad.

Throw GYConversionException if conversion fails.

datumHashFromPlutus' ∷ MonadError GYTxMonadException m ⇒ DatumHash → m GYDatumHash

Convert a DatumHash to GYDatumHash in GYTxMonad.

Throw GYConversionException if conversion fails.

pubKeyHashFromPlutus' ∷ MonadError GYTxMonadException m ⇒ PubKeyHash → m GYPubKeyHash

Convert a PubKeyHash to GYPubKeyHash in GYTxMonad.

Throw GYConversionException if conversion fails.

advanceSlot' ∷ MonadError GYTxMonadException m ⇒ GYSlot → Natural → m GYSlot

Advance GYSlot forward in GYTxMonad. If slot value overflows, throw GYSlotOverflowException.

utxosDatums ∷ ∀ m a. (GYTxQueryMonad m, FromData a) ⇒ GYUTxOs → m (Map GYTxOutRef (GYAddress, GYValue, a))

utxoDatum ∷ (GYTxQueryMonad m, FromData a) ⇒ GYUTxO → m (Either GYQueryDatumError (GYAddress, GYValue, a))

utxoDatumHushed ∷ (GYTxQueryMonad m, FromData a) ⇒ GYUTxO → m (Maybe (GYAddress, GYValue, a))

utxoDatum' ∷ (GYTxQueryMonad m, FromData a) ⇒ GYUTxO → m (GYAddress, GYValue, a)

Version of utxoDatum that throws GYTxMonadException.

mustHaveInput ∷ GYTxIn v → GYTxSkeleton v

mustHaveRefInput ∷ VersionIsGreaterOrEqual v PlutusV2 ⇒ GYTxOutRef → GYTxSkeleton v

mustHaveOutput ∷ GYTxOut v → GYTxSkeleton v

mustHaveOptionalOutput ∷ Maybe (GYTxOut v) → GYTxSkeleton v

mustMint ∷ GYMintingPolicy u → GYRedeemer → GYTokenName → Integer → GYTxSkeleton v

mustBeSignedBy ∷ GYPubKeyHash → GYTxSkeleton v

isInvalidBefore ∷ GYSlot → GYTxSkeleton v

isInvalidAfter ∷ GYSlot → GYTxSkeleton v

gyLogDebug' ∷ GYTxQueryMonad m ⇒ GYLogNamespace → String → m ()

gyLogInfo' ∷ GYTxQueryMonad m ⇒ GYLogNamespace → String → m ()

gyLogWarning' ∷ GYTxQueryMonad m ⇒ GYLogNamespace → String → m ()

gyLogError' ∷ GYTxQueryMonad m ⇒ GYLogNamespace → String → m ()

skeletonToRefScriptsORefs ∷ GYTxSkeleton v → [GYTxOutRef]

Given a skeleton, returns a list of reference to reference script UTxOs which are present as witness.