GeniusYield.TxBuilder.Errors

Description

Documentation

data PlutusToCardanoError

PlutusToCardanoError is raised when using Plutus to Cardano API type conversion functions from plutus-ledger. It is a focused version of Ledger.Tx.CardanoAPI.ToCardanoError.

Constructors

DeserialiseRawBytesError
Deserialization failed; tag indicates the type being deserialized.

Fields

ptceTag ∷ Text

StakePtrAddressUnsupported Address
Raised when trying to convert a stake ptr plutus address.

UnknownPlutusToCardanoError
Wildcard unhandled constructors; shouldn't happen usually.

Fields

ptceTag ∷ Text

data GYConversionError

GYConversionErrors may be raised during type conversions.

Constructors

GYNotPubKeyAddress !GYAddress
An address was expected to contain a pub key hash, but it did not.

GYNotScriptAddress !GYAddress
An address was expected to contain a script hash, but it did not.

GYInvalidPlutusValue !GYFromPlutusValueError !Value
Raised during Plutus Value to GYValue conversion.

GYInvalidPlutusAsset !GYFromPlutusValueError
Raised during Plutus asset to GY asset conversion.

GYInvalidAddressText !Text
Raised when trying to parse Text into GYAddress.

GYEraSummariesToSlotConfigError !Text
Raised when trying to convert EraHistory to GYSlotConfig.

GYLedgerToCardanoError !PlutusToCardanoError
Errors raised during plutus-ledger -> cardano api type conversion.

GYInvalidAssetClass !Text
Errors raised by "GeniusYield.Types.Value.parseAssetClassCore" and similar.

GYInvalidSlot !Integer
Errors caused by "GeniusYield.Types.Slot.slotFromInteger" resulting in Nothing.

data GYQueryUTxOError

GYQueryUTxOErrors may be raised during utxo related queries.

Constructors

GYNoUtxosAtAddress ![GYAddress]
An address was queried for one or more UTxOs but none were found.

GYNoUtxoAtRef !GYTxOutRef
No UTxO exists at given ref.

data GYQueryDatumError

GYQueryDatumError may be raised during fetching and parsing datums.

Constructors

GYNoDatumForHash !GYDatumHash
No datum found for given hash.

GYInvalidDatum !GYDatum
Datum parsing failed.

GYNoDatumHash !GYUTxO
No datum hash at utxo.

data GYTxMonadException ∷ Type where

Exceptions raised within the GYTxMonad computation.

This includes exceptions raised within the contract itself. It does not include:

Exceptions that may be raised by the provider.

Exceptions raised during transaction building/balancing.

Other wildcard exceptions raised within IO.

Constructors

GYConversionException ∷ GYConversionError → GYTxMonadException
Errors encountered during type conversions.

GYQueryUTxOException ∷ GYQueryUTxOError → GYTxMonadException
Errors encountered during utxo related queries.

GYNoSuitableCollateralException
Raised when no suitable collateral of at least tmeMinLovelace amount is found at tmeAddress.

Fields

∷ { tmeMinLovelace ∷ Natural

, tmeAddress ∷ GYAddress

} → GYTxMonadException

GYSlotOverflowException
Raised if tmeSlot value overflows when advancing it by tmeAdvanceAmount.

Fields

∷ { tmeSlot ∷ GYSlot

, tmeAdvanceAmount ∷ Natural

} → GYTxMonadException

GYTimeUnderflowException ∷ SystemStart → GYTime → GYTxMonadException
Raised during time -> slot conversion, if given timestamp is before known system start.

GYQueryDatumException ∷ GYQueryDatumError → GYTxMonadException
Raised during fetching/parsing datums.

GYApplicationException ∷ (Exception e, IsGYApiError e) ⇒ e → GYTxMonadException
Wildcard user application specific errors. This is the "plug-in" point where an application using the GY framework, can raise its own protocol specific errors within GYTxMonad.

throwAppError ∷ (IsGYApiError e, Exception e, MonadError GYTxMonadException m) ⇒ e → m a

Throw an application specific exception (GYApplicationException) within GYTxMonad.