GeniusYield.TxBuilder.Random

Description

Documentation

class Monad m ⇒ MonadRandom (m ∷ Type → Type) where

Methods

getRandomR ∷ Random a ⇒ (a, a) → m a

getRandom ∷ Random a ⇒ m a

getRandomRs ∷ Random a ⇒ (a, a) → m [a]

getRandoms ∷ Random a ⇒ m [a]

withRandomness ∷ GYTxMonad m ⇒ (∀ n. (GYTxMonad n, MonadRandom n) ⇒ n a) → m a

Convert a GYTxMonad-computation using randomness into a deterministic one by using one's own public key hash as seed for the random number generator.