GeniusYield.Types.Ledger

Description

Documentation

data PlutusToCardanoError

PlutusToCardanoError is raised when using Plutus to Cardano API type conversion functions from plutus-ledger. It is a focused version of Ledger.Tx.CardanoAPI.ToCardanoError.

Constructors

DeserialiseRawBytesError
Deserialization failed; tag indicates the type being deserialized.

Fields

ptceTag ∷ Text

StakePtrAddressUnsupported Address
Raised when trying to convert a stake ptr plutus address.

UnknownPlutusToCardanoError
Wildcard unhandled constructors; shouldn't happen usually.

Fields

ptceTag ∷ Text