GeniusYield.Types.Time

Description

Documentation

class FormatTime t where

Methods

formatCharacter ∷ Bool → Char → Maybe (FormatOptions → t → String)

Since: time-1.9.1

class ParseTime t where

The class of types which can be parsed given a UNIX-style time format string.

Minimal complete definition

parseTimeSpecifier, buildTime

Methods

substituteTimeSpecifier ∷ proxy t → TimeLocale → Char → Maybe String

Since: time-1.9.1

parseTimeSpecifier ∷ proxy t → TimeLocale → Maybe ParseNumericPadding → Char → ReadP String

Get the string corresponding to the given format specifier.

Since: time-1.9.1

buildTime

Arguments

∷ TimeLocale
The time locale.

→ [(Char, String)]
Pairs of format characters and the corresponding part of the input.

→ Maybe t

Builds a time value from a parsed input string. If the input does not include all the information needed to construct a complete value, any missing parts should be taken from 1970-01-01 00:00:00 +0000 (which was a Thursday). In the absence of %C or %Y, century is 1969 - 2068.

Since: time-1.9.1

gyIso8601Show ∷ GYTime → String

>>> gyIso8601Show (timeFromPlutus 0)
"1970-01-01T00:00:00Z"

gyIso8601ParseM ∷ MonadFail m ⇒ String → m GYTime

>>> gyIso8601ParseM @Maybe "1970-01-01T00:00:33.333Z"
Just (GYTime 33.333s)

>>> gyIso8601ParseM @Maybe "1970-01-01T00:00:33.333"
Nothing

data GYTime

getCurrentGYTime ∷ IO GYTime

addSeconds ∷ GYTime → Rational → GYTime

timeFromPlutus ∷ POSIXTime → GYTime

>>> timeFromPlutus 12345
GYTime 12.345s

timeToPlutus ∷ GYTime → POSIXTime

>>> timeToPlutus $ timeFromPlutus 31415
POSIXTime {getPOSIXTime = 31415}

timeToPOSIX ∷ GYTime → POSIXTime

>>> timeToPOSIX (timeFromPOSIX 12346)
12346s

timeFromPOSIX ∷ POSIXTime → GYTime

>>> timeFromPOSIX 12346
GYTime 12346s