GeniusYield.Types.TxBody

Description

Transaction body

data GYTxBody

Transaction body: the part which is then signed.

Conversions

txBodyFromApi ∷ TxBody BabbageEra → GYTxBody

txBodyToApi ∷ GYTxBody → TxBody BabbageEra

Transaction creation

signTx ∷ ToShelleyWitnessSigningKey a ⇒ GYTxBody → [a] → GYTx

Sign a transaction body with (potentially) multiple keys.

unsignedTx ∷ GYTxBody → GYTx

Create an unsigned transaction from the body.

Functions

txBodyFee ∷ GYTxBody → Integer

Return the fees in lovelace.

txBodyFeeValue ∷ GYTxBody → GYValue

Return the fees as GYValue.

txBodyUTxOs ∷ GYTxBody → GYUTxOs

Return utxos created by tx (body).

txBodyTxIns ∷ GYTxBody → [GYTxOutRef]

Returns the GYTxOutRef consumed by the tx.

txBodyTxInsReference ∷ GYTxBody → [GYTxOutRef]

Returns the GYTxOutRef for the reference inputs present in the tx.

txBodyTxId ∷ GYTxBody → GYTxId

Returns the GYTxId of the given GYTxBody.

txBodyToApiTxBodyContent ∷ GYTxBody → TxBodyContent ViewTx BabbageEra

txBodyMintValue ∷ GYTxBody → GYValue

Returns the mint GYValue of the given GYTxBody.

txBodyValidityRange ∷ GYTxBody → (Maybe GYSlot, Maybe GYSlot)

Returns the validity range of the given GYTxBody.

txBodyCollateral ∷ GYTxBody → Set GYTxOutRef

Returns the set of GYTxOutRef used as collateral in the given GYTxBody.

txBodyCollateralReturnOutput ∷ GYTxBody → TxReturnCollateral CtxTx BabbageEra

txBodyTotalCollateralLovelace ∷ GYTxBody → Natural

Returns the lovelace sum of all collateral used in the given GYTxBody.

getTxBody ∷ GYTx → GYTxBody

Returns the GYTxBody of the given GYTx.