GeniusYield.Types.TxIn

Description

Documentation

data GYTxIn v

Transaction input:

an UTxO

non-key witness for script utxos

The parameter v indicates the minimum version of scripts allowed as inputs in the transaction.

Constructors

GYTxIn

Fields

gyTxInTxOutRef ∷ !GYTxOutRef

gyTxInWitness ∷ !(GYTxInWitness v)

data GYInScript (u ∷ PlutusVersion) where

Constructors

GYInScript ∷ v `VersionIsGreaterOrEqual` u ⇒ GYValidator v → GYInScript u
VersionIsGreaterOrEqual restricts which version validators can be used in this transaction.

GYInReference ∷ !GYTxOutRef → !(GYScript PlutusV2) → GYInScript PlutusV2
Reference inputs can be only used in V2 transactions.

inScriptVersion ∷ GYInScript v → PlutusVersion

Returns the PlutusVersion of the given GYInScript.

data GYTxInWitness v

Represents witness type and associated information for tx inputs.

Constructors

GYTxInWitnessKey
Key witness without datum.

GYTxInWitnessScript !(GYInScript v) !GYDatum !GYRedeemer
Script witness with associated script, datum, and redeemer.

txInToApi

Arguments

∷ Bool
does corresponding utxo contains inline datum?

→ GYTxIn v

→ (TxIn, BuildTxWith BuildTx (Witness WitCtxTxIn BabbageEra))

Note: TxIns type synonym is not exported: https://github.com/input-output-hk/cardano-node/issues/3732