GeniusYield.Types.TxOut

Description

Documentation

data GYTxOut (v ∷ PlutusVersion)

Transaction output.

The parameter v indicates the minimum version of scripts allowed as inputs in the transaction.

Constructors

GYTxOut

Fields

gyTxOutAddress ∷ !GYAddress

gyTxOutValue ∷ !GYValue

gyTxOutDatum ∷ !(Maybe (GYDatum, GYTxOutUseInlineDatum v))

The Boolean indicates whether to use inline datums or not. May be overridden by a flag to txOutToApi.

gyTxOutRefS ∷ !(Maybe (GYScript PlutusV2))

data GYTxOutUseInlineDatum (v ∷ PlutusVersion) where

Constructors

GYTxOutUseInlineDatum ∷ GYTxOutUseInlineDatum PlutusV2

GYTxOutDontUseInlineDatum ∷ GYTxOutUseInlineDatum v

gyTxOutDatumL ∷ Traversal (GYTxOut v) (GYTxOut u) (GYTxOutUseInlineDatum v) (GYTxOutUseInlineDatum u)

Whether to use inline datum in this transaction output

mkGYTxOut ∷ GYAddress → GYValue → GYDatum → GYTxOut v

The most common variant: create a GYTxOut from address, value and datum

mkGYTxOutNoDatum ∷ GYAddress → GYValue → GYTxOut v

Same as mkGYTxOut but without a datum.

txOutToApi

Arguments

∷ Bool
True to add inline datum, otherwise use TxOutDatumInTx.

→ GYTxOut v

→ TxOut CtxTx BabbageEra