GeniusYield.Types.UTxO

Description

Documentation

data GYUTxO

An unspent transaction output.

Constructors

GYUTxO

Fields

utxoRef ∷ !GYTxOutRef

utxoAddress ∷ !GYAddress

utxoValue ∷ !GYValue

utxoOutDatum ∷ !GYOutDatum

utxoRefScript ∷ !(Maybe (Some GYScript))

utxoFromApi ∷ TxIn → TxOut CtxTx BabbageEra → GYUTxO

utxoFromApi' ∷ TxIn → TxOut CtxUTxO era → GYUTxO

utxoToPlutus ∷ GYUTxO → TxOut

data GYUTxOs

A set of unspent transaction outputs.

Actually a map from unspent transaction outputs to address, value and datum hash.

utxosSize ∷ GYUTxOs → Int

Number of UTxOs within given GYUTxOs.

utxosFromApi ∷ UTxO era → GYUTxOs

utxosToApi ∷ GYUTxOs → UTxO BabbageEra

utxosRemoveTxOutRef ∷ GYTxOutRef → GYUTxOs → GYUTxOs

Remove particular GYTxOutRef from GYUTxOs.

Used to remove collateral, so we don't use it in transactions.

utxosRemoveTxOutRefs ∷ Set GYTxOutRef → GYUTxOs → GYUTxOs

Remove several GYTxOutRefs from GYUTxOs.

utxosRemoveRefScripts ∷ GYUTxOs → GYUTxOs

Remove UTxOs containing reference scripts inside them from GYUTxOs.

utxosLookup ∷ GYTxOutRef → GYUTxOs → Maybe GYUTxO

Lookup a UTxO given a ref.

someTxOutRef ∷ GYUTxOs → Maybe (GYTxOutRef, GYUTxOs)

Get some output reference from GYUTxOs.

Used to pick an input for minting, or selecting collateral (in tests).

randomTxOutRef ∷ MonadRandom m ⇒ GYUTxOs → m (Maybe (GYTxOutRef, GYUTxOs))

Get a random output reference from GYUTxOs.

UTxO datums

data GYOutDatum

The datum contained within a transaction output.

Constructors

GYOutDatumNone
The output has no datum.

GYOutDatumHash !GYDatumHash
The output contains a datum hash, the associated datum may or may not be included in the tx.

GYOutDatumInline !GYDatum
The output contains an inline datum (i.e datum within the output itself).

isInlineDatum ∷ GYOutDatum → Bool

outDatumToPlutus ∷ GYOutDatum → OutputDatum

Filter and map

filterUTxOs ∷ (GYUTxO → Bool) → GYUTxOs → GYUTxOs

Filter GYUTxOs with a predicate on GYUTxO.

mapMaybeUTxOs ∷ (GYUTxO → Maybe a) → GYUTxOs → Map GYTxOutRef a

Map & filter GYUTxOs contents.

mapUTxOs ∷ (GYUTxO → a) → GYUTxOs → Map GYTxOutRef a

Map GYUTxOs contents.

witherUTxOs ∷ Applicative f ⇒ (GYUTxO → f (Maybe a)) → GYUTxOs → f (Map GYTxOutRef a)

Applicative version of mapMaybeUTxOs.

List conversions

utxosFromList ∷ [GYUTxO] → GYUTxOs

Returns a GYUTxOs from a given list of GYUTxOs.

utxosFromUTxO ∷ GYUTxO → GYUTxOs

Returns a GYUTxOs from a single GYUTxO.

utxosToList ∷ GYUTxOs → [GYUTxO]

Returns a list of GYUTxOs from a given GYUTxOs.

Folds

foldlUTxOs' ∷ ∀ a. (a → GYUTxO → a) → a → GYUTxOs → a

Fold operation over a GYUTxOs.

foldMapUTxOs ∷ Monoid m ⇒ (GYUTxO → m) → GYUTxOs → m

FoldMap operation over a GYUTxOs.

forUTxOs_ ∷ ∀ f a. Applicative f ⇒ GYUTxOs → (GYUTxO → f a) → f ()

foldMUTxOs ∷ ∀ m a. Monad m ⇒ (a → GYUTxO → m a) → a → GYUTxOs → m a

Extract refs

utxosRefs ∷ GYUTxOs → [GYTxOutRef]

Returns a list of all GYTxOutRefs inside a given GYUTxOs.