GeniusYield.Types.Value

Description

Value

data GYValue

Value: a (total) map from asset classes (GYAssetClass) to amount (Integer).

valueMake ∷ Map GYAssetClass Integer → GYValue

valueToPlutus ∷ GYValue → Value

Converts a GYValue to a Plutus Value

valueFromPlutus ∷ Value → Either GYFromPlutusValueError GYValue

Converts a Plutus Value to a GYValue. Returns Left GYFromPlutusValueError if it fails.

valueToApi ∷ GYValue → Value

Convert a GYValue to a Cardano Api Value

valueFromApi ∷ Value → GYValue

Convert a Cardano Api Value to a GYValue

valueSingleton ∷ GYAssetClass → Integer → GYValue

Returns a GYValue containing only the given GYAssetClass with the given amount.

>>> valueSingleton (GYToken "ff80aaaf03a273b8f5c558168dc0e2377eea810badbae6eceefc14ef" "GOLD") 100
valueFromList [(GYToken "ff80aaaf03a273b8f5c558168dc0e2377eea810badbae6eceefc14ef" "GOLD",100)]

valueFromList ∷ [(GYAssetClass, Integer)] → GYValue

Create GYValue from a list of asset class and amount. Duplicates are merged.

valueToList ∷ GYValue → [(GYAssetClass, Integer)]

Returns a list of all GYAssetClass and amount pairs contained in a given GYValue.

valueToMap ∷ GYValue → Map GYAssetClass Integer

Returns a map from GYAssetClass to their amount contained in a given GYValue.

valueMap ∷ (GYAssetClass → Integer → Integer) → GYValue → GYValue

Map operation over the GYValues amounts.

valueTotalAssets ∷ GYValue → Int

Returns the total count of assets in a given GYValue

valueInsert ∷ GYAssetClass → Integer → GYValue → GYValue

Insert an asset with its quantity.

valueAdjust ∷ (Integer → Integer) → GYAssetClass → GYValue → GYValue

Adjust the amount of a given GYAssetClass in the given GYValue.

valueFromLovelace ∷ Integer → GYValue

>>> valueFromLovelace 0
valueFromList []

>>> valueFromLovelace 100
valueFromList [(GYLovelace,100)]

valueFromApiTxOutValue ∷ TxOutValue era → GYValue

valueToApiTxOutValue ∷ GYValue → TxOutValue BabbageEra

valueAssets ∷ GYValue → Set GYAssetClass

Set of assets within a GYValue in non-zero quantities.

Arithmetic

valueMinus ∷ GYValue → GYValue → GYValue

Substracts the second GYValue from the first one. AssetClass by AssetClass.

valueNegate ∷ GYValue → GYValue

Returns the given GYValue with all amounts negated.

valuePositive ∷ GYValue → Bool

Checks if all amounts of the given GYValue are positive ( > 0 ).

valueNonNegative ∷ GYValue → Bool

Checks if all amounts of the given GYValue are not negative ( >= 0 ).

valueGreaterOrEqual ∷ GYValue → GYValue → Bool

Checks if all amounts of the first GYValue are greater or equal to the second GYValue.

valueGreater ∷ GYValue → GYValue → Bool

Checks if all amounts of the first GYValue are greater than the second GYValue.

Unions & Intersections

valueUnionWith ∷ (Integer → Integer → Integer) → GYValue → GYValue → GYValue

Combine two values with given function.

valueIntersection ∷ GYValue → GYValue → GYValue

Left biased intersection of two GYValues.

Persist only the assets in the first value that also exist in the second.

valueIntersectionWith ∷ (Integer → Integer → Integer) → GYValue → GYValue → GYValue

Intersection of two GYValues with a combining function.

Lookup

valueAssetClass ∷ GYValue → GYAssetClass → Integer

Returns the amount of a GYAssetClass contained in the given GYValue.

Splitting

valueSplitAda ∷ GYValue → (Integer, GYValue)

Splits a GYValue into the lovelace amount and the rest of it's components.

>>> valueSplitAda $ valueFromLovelace 100
(100,valueFromList [])

>>> valueSplitAda $ valueFromList [(GYLovelace, 100), (GYToken "ff80aaaf03a273b8f5c558168dc0e2377eea810badbae6eceefc14ef" "GOLD",101)]
(100,valueFromList [(GYToken "ff80aaaf03a273b8f5c558168dc0e2377eea810badbae6eceefc14ef" "GOLD",101)])

valueSplitSign ∷ GYValue → (GYValue, GYValue)

Split a GYValue into its positive and negative components. The first element of the pair is the positive components of the value. The second element is the negative component.

>>> valueSplitSign $ valueFromList [(GYLovelace,22),(GYToken "ff80aaaf03a273b8f5c558168dc0e2377eea810badbae6eceefc14ef" "GOLD",-10)]
(valueFromList [(GYLovelace,22)],valueFromList [(GYToken "ff80aaaf03a273b8f5c558168dc0e2377eea810badbae6eceefc14ef" "GOLD",10)])

Predicates

isEmptyValue ∷ GYValue → Bool

Checks if the given GYValue is empty

>>> isEmptyValue mempty
True

>>> isEmptyValue $ valueFromLovelace 100
False

>>> isEmptyValue $ valueMinus (valueFromLovelace 100) (valueFromLovelace 100)
True

valueVerifyNonNegative ∷ GYValue → Maybe (Map GYAssetClass Natural)

Verify the value only consists of positive amounts, returning a map containing naturals as a result.

Debug

valueValid ∷ GYValue → Bool

Check the GYValue representation invariants.

Should always evaluate to True

Conversion errors

data GYFromPlutusValueError

Errors raised during Value -> GYValue conversion.

Constructors

GYTokenNameTooBig !TokenName
Length of the token name bytestring is more than 32.

GYInvalidPolicyId !CurrencySymbol
PolicyId deserialization failure.

Asset class

data GYAssetClass

Asset class. Either lovelace or minted token.

Constructors

GYLovelace

GYToken GYMintingPolicyId GYTokenName

assetClassToPlutus ∷ GYAssetClass → AssetClass

Converts a GYAssetClass into a Plutus AssetClass.

assetClassFromPlutus ∷ AssetClass → Either GYFromPlutusValueError GYAssetClass

Converts a Plutus AssetClass into a GYAssetClass. Returns Left GYFromPlutusValueError if it fails.

parseAssetClassWithSep ∷ Char → Text → Either String GYAssetClass

Parse hex encoded currency symbol and hex encoded token name separated by the given separator. >>> parseAssetClassWithSep . "lovelace" Right GYLovelace

>>> parseAssetClassWithSep '.' ""
Right GYLovelace

>>> parseAssetClassWithSep '.' "ff80aaaf03a273b8f5c558168dc0e2377eea810badbae6eceefc14ef.476f6c64"
Right (GYToken "ff80aaaf03a273b8f5c558168dc0e2377eea810badbae6eceefc14ef" "Gold")
>>> parseAssetClassWithSep '#' "ff80aaaf03a273b8f5c558168dc0e2377eea810badbae6eceefc14ef#476f6c64"
Right (GYToken "ff80aaaf03a273b8f5c558168dc0e2377eea810badbae6eceefc14ef" "Gold")
>>> parseAssetClassWithSep '#' "ff80aaaf03a273b8f5c558168dc0e2377eea810badbae6eceefc14ef#"
Right (GYToken "ff80aaaf03a273b8f5c558168dc0e2377eea810badbae6eceefc14ef" "")

parseAssetClassCore ∷ Char → (ByteString → Parser GYTokenName) → Text → Either String GYAssetClass

Token name

newtype GYTokenName

Token name is an arbitrary byte string up to 32 bytes long.

TODO: it's unclear whether it's an arbitrary byte string or UTF8 encoded text (which encoded byte form is 32 byte long at most). We treat it as an arbitrary string.

>>> LBS8.putStrLn $ Aeson.encode ("Gold" :: GYTokenName)
"476f6c64"

Constructors

GYTokenName ByteString

tokenNameToHex ∷ GYTokenName → Text

tokenNameFromBS ∷ ByteString → Maybe GYTokenName

tokenNameToPlutus ∷ GYTokenName → TokenName

tokenNameFromPlutus ∷ HasCallStack ⇒ TokenName → Maybe GYTokenName

Convert Plutus TokenName to GYTokenName.

tokenNameFromHex ∷ Text → Either Text GYTokenName

unsafeTokenNameFromHex ∷ Text → GYTokenName

makeAssetClass ∷ Text → Text → Either String GYAssetClass